perf(memory-wiki): skip link parsing for bracket-free pages
